Update README.md
At the beginning, we should import the required libraries:
    import cv2
    import numpy as np
    import random
For “random” it works with randomness, including generating random numbers, shuffling sequences, and making random choices from a list we can use this library. use this library for abstract shapes.
To have a canvas of 800X800 for color drawing of a doodle image with a white color.